% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FortCollins}
\alias{FortCollins}
\title{Fort Collins weather data}
\format{
A data frame with 36524 rows and four variables:
\describe{
\item{date}{\code{Date} of observation}
\item{precip}{daily cumulated precipitation (in mm); most are zero}
\item{maxTemp}{maximum daily temperature temperature (in Celcius)}
\item{minTemp}{minimum daily temperature (in Celcius)}
}
}
\source{
Colorado Climate Center at Colorado State University
}
\usage{
FortCollins
}
\description{
This data is extracted from the \code{extRemes} package
and consists of daily cumulated precipitation (in mm), minimum and maximum temperature (in degree Celcius).
The astute reader will notice that the precipitation series
(reported in inches according to the documentation) has been multiplied by 100.
}
\references{
Katz, R. W., Parlange, M. B. and Naveau, P. (2002) Statistics of extremes in hydrology. \emph{Advances in Water Resources}, \bold{25}, 1287–1304.
}
\seealso{
\code{\link[extRemes]{FCwx}}
}
\keyword{datasets}
