% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{returnsBanks}
\alias{returnsBanks}
\title{Daily returns for three major US banks}
\format{
A data frame with 5283 rows and the following four variables
\describe{
\item{date}{\code{Date} at which the return is calculated}
\item{retC}{returns for Chase}
\item{retBAC}{returns for Bank of America}
\item{retGS}{returns for Goldman Sachs}
}
}
\usage{
returnsBanks
}
\description{
This data frame contains daily returns for three major
banks, namely Chase (\code{C}), Bank of America (\code{BAC})
and Goldman Sachs (\code{GS}) for the period 2000-2020.
The data were extracted using `tidyquant` and computed from
adjusted returns, with the difference relative to the previous day
of operation, with \eqn{r_t=(Y_t-Y_{t-1})/Y_{t-1}}.
The \code{date} field contains the date
during which the return is calculated.
}
\keyword{datasets}
